Implemented garbage collection after each commit.
New directories are marked and sorted by location of dirent block.
The dirents are relocated toward the start of the pool.
The contents of the dirents are changed to where the dirs will go.
The dirs are relocated toward the start of the pool.
Any empty dirs are remapped to the first dir in the pool.

Signed-off-by: David Barr <david.barr@cordelta.com>