Simple validation script with hashification.
This script depends on SVK for fast checkout and update.
Revisions are walked at the root of the repository.
After each update new files are hardlinked to equivalent files.

When the final revision is reached, the repository is imported.
The revisions are then walked in reverse and git reset confirms.

Signed-off-by: David Barr <david.barr@cordelta.com>